src/functions/exp/requirements.txt
aiohttp==3.7.4.post0
astroid==2.11.5
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
black==22.3.0
CacheControl==0.12.11
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cloudevents==1.2.0
debugpy==1.6.0
decorator==5.1.1
deprecation==2.1.0
dill==0.3.4
discord==1.7.3
discord.py==1.7.3
distlib==0.3.4
entrypoints==0.4
executing==0.8.3
filelock==3.7.0
firebase-admin==4.5.1
Flask==2.1.2
functions-framework==3.0.0
google-api-core==1.31.5
google-api-python-client==2.47.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-pubsub==2.13.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
grpc-google-iam-v1==0.12.4
grpcio==1.46.1
grpcio-status==1.46.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
install==1.3.5
ipykernel==6.13.0
ipython==8.3.0
iso8601==1.0.2
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.1
jupyter-core==4.10.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pylint==2.13.9
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.0.0
requests==2.27.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tox==3.25.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
watchdog==1.0.2
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
